+++ scripts/systems/dnd4e.js
@@ -22,33 +22,45 @@ class Engine4e extends Engine {
     });
   }
 
-  canDetectHidden(visionSource, hiddenEffect, tgtToken) {
-    // Never gets called, neither do the patches for the v10 vision modes
-    // dead in the water
-    const source = visionSource.object?.actor;
-    const stealth = hiddenEffect.flags.stealthy?.hidden ?? (10 + tgtToken.actor.system.skills.stl.total);
-    const spotEffect = this.findSpotEffect(source);
-    const perception = spotEffect?.flags.stealthy?.spot ?? (10 + source.system.skills.prc.total);
+  canDetectHidden(visionSource, tgtToken, detectionMode) {
+    const stealthFlag = this.getStealthFlag(tgtToken);
+    if (!stealthFlag) return true;
 
-    return  perception > stealth;
+    const stealthValue = this.getStealthValue(stealthFlag);
+    const perceptionFlag = this.getPerceptionFlag(visionSource.object);
+    const perceptionValue = this.getPerceptionValue(perceptionFlag);
+
+    return perceptionValue > stealthValue;
   }
 
-  getHiddenFlagAndValue(actor, effect) {
-    const value = effect.flags.stealthy?.hidden ?? (10 + actor.system.skills.stl.total);
-    return { flag: { hidden: value }, value };
+  getStealthFlag(token) {
+    let flag = super.getStealthFlag(token);
+    if (flag && flag.stealth === undefined)
+      flag.stealth = 10 + token.actor.system.skills.stl.total;
+    return flag;
   }
 
-  getSpotFlagAndValue(actor, effect) {
-    const value = effect?.flags?.stealthy?.spot ?? (10 + actor.system.skills.prc.total);
-    return { flag: { spot: value }, value };
+  getPerceptionFlag(token) {
+    const flag = super.getPerceptionFlag(token);
+    if (flag) return flag;
+    return {
+      token,
+      passive: true,
+      perception: 10 + token.actor.system.skills.prc.total
+    };
   }
 
   async rollPerception(message, options, id) {
     Stealthy.log('rollPerception', { message, options, id });
+    if (!stealthy.bankingPerception) return;
 
     const token = canvas.tokens.get(message.speaker.token);
-    const actor = token.actor;
-    await this.updateOrCreateSpotEffect(actor, { spot: message.rolls[0].total });
+    if (stealthy.perceptionToActor) {
+      const actor = token.actor;
+      await this.updateOrCreateSpotEffect(actor, { perception: message.rolls[0].total });
+    } else {
+      await this.putRollOnToken(token, 'perception', message.rolls[0].total);
+    }
 
     super.rollPerception();
   }
@@ -57,8 +69,12 @@ class Engine4e extends Engine {
     Stealthy.log('rollStealth', { message, options, id });
 
     const token = canvas.tokens.get(message.speaker.token);
-    const actor = token.actor;
-    await this.updateOrCreateHiddenEffect(actor, { hidden: message.rolls[0].total });
+    if (stealthy.stealthToActor) {
+      const actor = token.actor;
+      await this.updateOrCreateHiddenEffect(actor, { stealth: message.rolls[0].total });
+    } else {
+      await this.putRollOnToken(token, 'stealth', message.rolls[0].total);
+    }
 
     super.rollStealth();
   }