+++ scripts/systems/dnd5e.js
@@ -23,6 +23,33 @@ class Engine5e extends Engine {
     });
 
     Hooks.once('setup', () => {
+      game.settings.register(Stealthy.MODULE_ID, 'perceptionDisadvantage', {
+        name: game.i18n.localize("stealthy.dnd5e.perceptionDisadvantage.name"),
+        hint: game.i18n.localize("stealthy.dnd5e.perceptionDisadvantage.hint"),
+        scope: 'world',
+        config: true,
+        type: Boolean,
+        default: true,
+      });
+
+      game.settings.register(Stealthy.MODULE_ID, 'stealthKey', {
+        name: game.i18n.localize("stealthy.dnd5e.stealthKey.name"),
+        hint: game.i18n.localize("stealthy.dnd5e.stealthKey.hint"),
+        scope: 'world',
+        config: true,
+        type: String,
+        default: 'ste'
+      });
+
+      game.settings.register(Stealthy.MODULE_ID, 'perceptionKey', {
+        name: game.i18n.localize("stealthy.dnd5e.perceptionKey.name"),
+        hint: game.i18n.localize("stealthy.dnd5e.perceptionKey.hint"),
+        scope: 'world',
+        config: true,
+        type: String,
+        default: 'prc'
+      });
+
       game.settings.register(Stealthy.MODULE_ID, 'ignorePassiveFloor', {
         name: game.i18n.localize("stealthy.dnd5e.ignorePassiveFloor.name"),
         hint: game.i18n.localize("stealthy.dnd5e.ignorePassiveFloor.hint"),
@@ -44,71 +71,13 @@ class Engine5e extends Engine {
         },
         default: 'inCombat'
       });
-
-      const tlcActive = game.modules.get("tokenlightcondition")?.active;
-
-      game.settings.register(Stealthy.MODULE_ID, 'tokenLighting', {
-        name: game.i18n.localize("stealthy.dnd5e.tokenLighting.name"),
-        hint: game.i18n.localize("stealthy.dnd5e.tokenLighting.hint"),
-        scope: 'world',
-        config: tlcActive,
-        type: Boolean,
-        default: false,
-      });
-
-      game.settings.register(Stealthy.MODULE_ID, 'spotPair', {
-        name: game.i18n.localize("stealthy.dnd5e.spotPair.name"),
-        hint: game.i18n.localize("stealthy.dnd5e.spotPair.hint"),
-        scope: 'world',
-        config: tlcActive,
-        type: Boolean,
-        default: false,
-      });
-
-      if (tlcActive) {
-        const v10 = Math.floor(game.version) < 11;
-        game.settings.register(Stealthy.MODULE_ID, 'darkLabel', {
-          name: game.i18n.localize("stealthy.dnd5e.dark.key"),
-          scope: 'world',
-          requiresReload: true,
-          config: true,
-          type: String,
-          default: v10 ? 'stealthy.dnd5e.dark.label' : 'stealthy.dnd5e.dark.name',
-        });
-
-        game.settings.register(Stealthy.MODULE_ID, 'dimLabel', {
-          name: game.i18n.localize("stealthy.dnd5e.dim.key"),
-          scope: 'world',
-          requiresReload: true,
-          config: true,
-          type: String,
-          default: v10 ? 'stealthy.dnd5e.dim.label' : 'stealthy.dnd5e.dim.name',
-        });
-
-        this.dimLabel = game.i18n.localize(game.settings.get(Stealthy.MODULE_ID, 'dimLabel'));
-        this.darkLabel = game.i18n.localize(game.settings.get(Stealthy.MODULE_ID, 'darkLabel'));
-        Stealthy.log(`dimLabel='${this.dimLabel}', darkLabel='${this.darkLabel}'`);
-
-        Hooks.on('renderSettingsConfig', (app, html, data) => {
-          $('<div>').addClass('form-group group-header')
-            .html('Token Lighting')
-            .insertBefore($('[name="stealthy.tokenLighting"]')
-              .parents('div.form-group:first'));
-        });
-      }
-      else {
-        Hooks.once('ready', () => {
-          game.settings.set(Stealthy.MODULE_ID, 'tokenLighting', false);
-          game.settings.set(Stealthy.MODULE_ID, 'spotPair', false);
-        });
-      }
     });
 
     Hooks.on('dnd5e.rollSkill', async (actor, roll, skill) => {
-      if (skill === 'ste') {
+      if (skill === game.settings.get(Stealthy.MODULE_ID, 'stealthKey')) {
         await this.rollStealth(actor, roll);
       }
-      else if (skill === 'prc') {
+      else if (skill === game.settings.get(Stealthy.MODULE_ID, 'perceptionKey')) {
         await this.rollPerception(actor, roll);
       }
     });
@@ -116,7 +85,7 @@ class Engine5e extends Engine {
     Hooks.on('renderSettingsConfig', (app, html, data) => {
       $('<div>').addClass('form-group group-header')
         .html(game.i18n.localize("stealthy.dnd5e.name"))
-        .insertBefore($('[name="stealthy.ignorePassiveFloor"]')
+        .insertBefore($('[name="stealthy.perceptionDisadvantage"]')
           .parents('div.form-group:first'));
     });
   }
@@ -166,78 +135,81 @@ class Engine5e extends Engine {
   }
 
   static LIGHT_LABELS = ['dark', 'dim', 'bright', 'bright'];
+  static EXPOSURE = { dark: 0, dim: 1, bright: 2 };
 
-  canDetectHidden(visionSource, hiddenEffect, tgtToken, detectionMode) {
+  canDetectHidden(visionSource, tgtToken, detectionMode) {
+    const stealthFlag = this.getStealthFlag(tgtToken);
+    if (!stealthFlag) return true;
+
     const srcToken = visionSource.object.document;
     const source = srcToken?.actor;
-    const stealth = hiddenEffect.flags.stealthy?.hidden ?? target.actor.system.skills.ste.passive;
-    const spotEffect = this.findSpotEffect(source);
+    const perceptionFlag = this.getPerceptionFlag(srcToken);
 
     // active perception loses ties, passive perception wins ties to simulate the
     // idea that active skills need to win outright to change the status quo. Passive
     // perception means that stealth is being the active skill.
-    const spotPair = spotEffect?.flags.stealthy?.spot;
-    let perception;
-
-    if (game.settings.get(Stealthy.MODULE_ID, 'tokenLighting')) {
-      perception = this.adjustForLightingConditions(spotPair, visionSource, source, tgtToken.actor, detectionMode);
+    const valuePair = perceptionFlag?.perception;
+    let perceptionValue;
+    if (game.settings.get(Stealthy.MODULE_ID, 'perceptionDisadvantage')) {
+      perceptionValue = this.adjustForLightingConditions(valuePair, visionSource, source, tgtToken, detectionMode);
     }
     else {
-      perception = this.adjustForDefaultConditions(spotPair, visionSource, source, tgtToken.actor, detectionMode);
+      perceptionValue = this.adjustForDefaultConditions(valuePair, visionSource, source, tgtToken, detectionMode);
     }
 
-    return perception > stealth;
+    const stealthValue = this.getStealthValue(stealthFlag);
+    Stealthy.logIfDebug('canDetectHidden', { stealthFlag: stealthFlag, stealth: stealthValue, perceptionFlag: perceptionFlag, perception: perceptionValue });
+    return perceptionValue > stealthValue;
   }
 
-  makeSpotEffectMaker(label) {
+  makeSpotEffectMaker(name) {
     return (flag, source) => {
-      let effect = super.makeSpotEffectMaker(label)(flag, source);
+      let effect = super.makeSpotEffectMaker(name)(flag, source);
       if (game.combat) effect.duration = { turns: 1, seconds: 6 };
       return effect;
     };
   }
 
-  getHiddenFlagAndValue(actor, effect) {
-    const value = effect?.flags.stealthy?.hidden ?? actor.system.skills.ste.passive;
+  getStealthFlag(token) {
+    let flag = super.getStealthFlag(token);
+    if (flag && flag.stealth === undefined)
+      flag.stealth = flag.token.actor.system.skills[game.settings.get(Stealthy.MODULE_ID, 'stealthKey')].passive;
+    return flag;
+  }
+
+  getPerceptionFlag(token) {
+    const flag = super.getPerceptionFlag(token);
+    if (flag) return flag;
+    const passive = token.actor.system.skills[game.settings.get(Stealthy.MODULE_ID, 'perceptionKey')].passive;
     return {
-      flag: { hidden: value },
-      value
+      token,
+      passive: true,
+      perception: {
+        normal: passive,
+        disadvantaged: passive - 5
+      }
     };
   }
 
-  getSpotFlagAndValue(actor, effect) {
-    let flag = { normal: undefined, disadvantaged: undefined };
-    const active = effect?.flags.stealthy?.spot?.normal ?? effect?.flags.stealthy?.spot;
-    if (active !== undefined) {
-      flag.normal = active;
-      flag.disadvantaged = effect.flags.stealthy?.spot?.disadvantaged ?? active - 5;
-    }
+  getPerceptionValue(flag) {
+    return flag?.perception?.normal ?? flag?.perception;
+  }
+
+  async setPerceptionValue(flag, value) {
+    if (value === undefined)
+      await super.setPerceptionValue(flag, value);
     else {
-      flag.normal = actor.system.skills.prc.passive;
-      flag.disadvantaged = Engine5e.GetPassivePerceptionWithDisadvantage(actor);
+      const pair = { normal: value, disadvantaged: value - 5 };
+      await super.setPerceptionValue(flag, pair);
     }
-    return {
-      flag: { spot: flag },
-      value: flag.normal
-    };
-  }
-
-  async setSpotValue(actor, effect, flag, value) {
-    const delta = value - flag.spot.normal;
-    flag.spot.normal = value;
-    flag.spot.disadvantaged += delta;
-    effect.flags.stealthy = flag;
-
-    await actor.updateEmbeddedDocuments('ActiveEffect', [effect]);
-    canvas.perception.update({ initializeVision: true }, true);
   }
 
   async rollPerception(actor, roll) {
-    if (!stealthy.activeSpot) return;
     Stealthy.log('Stealthy5e.rollPerception', { actor, roll });
+    if (!stealthy.bankingPerception) return;
 
     let perception = { normal: roll.total, disadvantaged: roll.total };
-    if (!roll.hasDisadvantage && game.settings.get(Stealthy.MODULE_ID, 'spotPair')) {
+    if (!roll.hasDisadvantage && game.settings.get(Stealthy.MODULE_ID, 'perceptionDisadvantage')) {
       const dice = roll.dice[0];
       if (roll.hasAdvantage) {
         const delta = dice.results[1].result - dice.results[0].result;
@@ -254,7 +226,17 @@ class Engine5e extends Engine {
       }
     }
 
-    await this.updateOrCreateSpotEffect(actor, { spot: perception });
+    if (!game.settings.get(Stealthy.MODULE_ID, 'ignorePassiveFloor')) {
+      const passivePrc = actor.system.skills[game.settings.get(Stealthy.MODULE_ID, 'perceptionKey')].passive;
+      perception.normal = Math.max(perception.normal, passivePrc);
+      perception.disadvantaged = Math.max(perception.disadvantaged, passivePrc - 5);
+    }
+
+    if (stealthy.perceptionToActor) {
+      await this.updateOrCreateSpotEffect(actor, { perception });
+    } else {
+      await this.putRollOnToken(actor, 'perception', perception);
+    }
 
     super.rollPerception();
   }
@@ -262,72 +244,57 @@ class Engine5e extends Engine {
   async rollStealth(actor, roll) {
     Stealthy.log('Stealthy5e.rollStealth', { actor, roll });
 
-    await this.updateOrCreateHiddenEffect(actor, { hidden: roll.total });
+    if (stealthy.stealthToActor) {
+      await this.updateOrCreateHiddenEffect(actor, { stealth: roll.total });
+    } else {
+      await this.putRollOnToken(actor, 'stealth', roll.total);
+    }
 
     super.rollStealth();
   }
 
   static GetPassivePerceptionWithDisadvantage(source) {
     // todo: don't apply -5 if already disadvantaged
-    return source.system.skills.prc.passive - 5;
+    return source.system.skills[game.settings.get(Stealthy.MODULE_ID, 'stealthKey')].passive - 5;
   }
 
-  adjustForDefaultConditions(spotPair, visionSource, source, target, detectionMode) {
-    const passivePrc = source?.system?.skills?.prc?.passive ?? -100;
-    let debugData = { passivePrc };
-    let perception = spotPair?.normal
-      ?? spotPair
+  adjustForDefaultConditions(perceptionPair, visionSource, source, tgtToken, detectionMode) {
+    const passivePrc = source?.system?.skills?.[game.settings.get(Stealthy.MODULE_ID, 'perceptionKey')]?.passive ?? -100;
+    let perception = perceptionPair?.normal
+      ?? perceptionPair
       ?? (passivePrc + 1);
-    debugData.perception = perception;
-    if (!game.settings.get(Stealthy.MODULE_ID, 'ignorePassiveFloor')) {
-      perception = Math.max(perception, passivePrc);
-      debugData.clampedPerception = perception;
-    }
-    Stealthy.logIfDebug('adjustForDefaultConditions', debugData);
     return perception;
   }
 
   // check target Token Lighting conditions via effects usage
   // look for effects that indicate Dim or Dark condition on the token
-  adjustForLightingConditions(spotPair, visionSource, source, target, detectionMode) {
-    let debugData = { spotPair };
+  adjustForLightingConditions(perceptionPair, visionSource, source, tgtToken, detectionMode) {
     let perception;
 
     // What light band are we told we sit in?
-    let lightBand = 2;
-    const v10 = Math.floor(game.version) < 11;
-    if (target?.effects.find(e => (v10 ? e.label : e.name) === this.darkLabel && !e.disabled)) { lightBand = 0; }
-    if (target?.effects.find(e => (v10 ? e.label : e.name) === this.dimLabel && !e.disabled)) { lightBand = 1; }
-    debugData.initialLightLevel = Engine5e.LIGHT_LABELS[lightBand];
+    let lightBand = Engine5e.EXPOSURE[this.getLightExposure(tgtToken)];
 
     // Adjust the light band based on conditions
     if (detectionMode) {
-      debugData.detectionMode = detectionMode;
       if (detectionMode === 'basicSight') {
         lightBand = lightBand + 1;
-        debugData.adjustedLightLevel = Engine5e.LIGHT_LABELS[lightBand];
       }
     }
     else {
-      debugData.id = visionSource.visionMode?.id;
       if (visionSource.visionMode?.id === 'darkvision') {
         lightBand = lightBand + 1;
-        debugData.adjustedLightLevel = Engine5e.LIGHT_LABELS[lightBand];
       }
     }
 
     // Extract the normal perception values from the source
-    const ignorePassiveFloor = game.settings.get(Stealthy.MODULE_ID, 'ignorePassiveFloor');
-    let active = spotPair?.normal ?? spotPair;
+    let active = perceptionPair?.normal ?? perceptionPair;
     let value;
-    const passivePrc = source?.system?.skills?.prc?.passive ?? -100;
+    const passivePrc = source?.system?.skills?.[game.settings.get(Stealthy.MODULE_ID, 'perceptionKey')]?.passive ?? -100;
     if (active !== undefined) {
       value = active;
-      debugData.active = value;
     }
     else {
       value = passivePrc;
-      debugData.passive = value;
     }
 
     // dark = fail, dim = disadvantage, bright = normal
@@ -336,22 +303,18 @@ class Engine5e extends Engine {
     }
     else if (lightBand === 1) {
       let passiveDisadv = Engine5e.GetPassivePerceptionWithDisadvantage(source);
-      debugData.passiveDisadv = passiveDisadv;
       if (active !== undefined) {
-        value = spotPair?.disadvantaged ?? value - 5;
-        debugData.activeDisadv = value;
+        value = perceptionPair?.disadvantaged ?? value - 5;
       }
       else {
         value = passiveDisadv;
       }
-      perception = (ignorePassiveFloor) ? value : Math.max(value, passiveDisadv);
+      perception = value;
     }
     else {
-      perception = (ignorePassiveFloor) ? value : Math.max(value, passivePrc);
+      perception = value;
     }
-    debugData.perception = perception;
 
-    Stealthy.logIfDebug('adjustForLightingConditions', debugData);
     return perception;
   }