+++ ChangeLog.md
@@ -1,4 +1,19 @@
+# v4.0.0
+* Big code changes, but games should have a seamless upgrade since no data migrations are required.
+* Stealthy now allows the choice of banking stealth and perception rolls to tokens or actors via effects.
+* Game setting defaults will bank perception rolls on the token and stealth rolls on an actor's effect.
+* 'Active Spot' control title changed to 'Bank perception rolls' and the control icon has changed to a piggy bank.
+* Token-banked rolls can be removed by clicking into the token HUD result and deleting its text. Actor-banked rolls are removed by deleting the effect. All banked perception rolls are still removed when toggling off the 'Bank perception rolls' token control.
+* New text hasn't been properly localized yet - I put english placeholders in the language json files.
+* Added client setting to add a lighting exposure icon to the token HUD.
+* Added a data schema setting to allow future data migration
+* Drop CUB support
+* Dnd5e: The 'perception disadvantage' setting enables tokens with stealth to now use the scene's lighting state to determine if they should apply disadvantage to the opposing perception check rather than looking up dim/dark effects on its actor
+* Dnd5e: Unless disabled, passive perception clamping of perception rolls now happens at time of roll rather than at comparison. This allows you to edit the banked perception roll to be less than the passive perception value if you choose.
+* Dnd5e: added settings to allow custom skills to substitute for stealth or perception
+
 # v3.18.1
+* FINAL V10-compatible version. Only V11+ in the future.
 * Update pt-BR.json (thanks Kharmans)
 
 # v3.18.0